Change the numbering order in the proposal list page
Fixes a minor annoyance in the proposal listing page, where the number
next to the proposal title is numbered from the top.
When a new proposal is submitted, the old proposals are shifted down and
the entire numbering is changed.

By using a revcounter in the forloop, we can show that the first talk
submitted is always displayed the number 1 even if it is pushed to the
bottom and the latest talk submitted will always be displayed the
highest number.